NOTE: from last Commit, ASP.NET was helpful enough to give us a return URL equal to where user started from so that if they manage to validate user i can shoot them back to where they tried to access
AUTHENTICATION & AUTHORIZATION
We will now have to tell ASP.NET who is logged in and to make our form actually work.  To fix the roles problem i created a Custom Role Provider ( FUNCTIONALITY built into ASP.NET) in main Web.Config file--> it tells ASP..NET what roles logged in user has
Basically create a class and tell ASP.NET what that class is
FormsAuthentication.SetAuthCookie(form.Username, true); <== Now that is how we tell ASP.NET that a person is who he says he is. ASP.NET handles all validation and encryption , all we have to check is make sure password is not blank. There's no username or password checking, just check to see if password/username is not blank.All we have to do is SetAuthCookie, give it a username and set it's persistance only
GetRolesForUser <==  Now that is how we tell ASP.NET that a particular person has access to a particular thing / role