*NOTE* - .ASPXUTH is an encrypted cookie that tells  ASP.NET who  is logged in
PROBLEM: Everone was allowed access to /admin/post because we are NOT checking to see what the usernames are! - and we want to prevent that. We will fix this in the Web.Config file with an IF Statement that if username == "nelson" to return "admin" otherwise return new empty string array ==> Other words if useris nelson then he is an admin and allowed access ottherwise he is nothing - Hacker Alert!